refactor: split frontend/src/components/search_palette.rs into submodules (#401)
The 801-line file is converted into a `search_palette/` subdirectory with:
- mod.rs    — re-exports the public SearchPaletteHost / PaletteOpen /
  use_palette_global_shortcut; carries the not(mobile) feature gate via the
  components/mod.rs declaration; hosts the trigger button + global hotkey
- overlay.rs — SpOverlay component (state machine, query signal, debounce,
  autofocus + async-sleep platform interop, SpFooter)
- results.rs — SpResultsList + per-hit row components
- keyboard.rs — make_keydown_handler closure builder
- model.rs   — shared FlatItem enum + small helpers (build_flat_items,
  is_selected, plural, facet_query) used by overlay/results/keyboard

Public API (the SearchPaletteHost + PaletteOpen + use_palette_global_shortcut
items consumed by components/mod.rs, lib.rs, and top_nav.rs) is unchanged.
No behaviour changes.

Closes #401
