--- frontend/src/components/search_palette.rs
@@ -1,801 +1,0 @@
-//! Command-palette / Spotlight-style search overlay. Replaces the inline
-//! nav search input with a trigger button in the top nav; clicking it (or
-//! pressing `⌘K` / `Ctrl+K`) opens a floating palette with debounced FTS5
-//! search and grouped results (Books, Authors, Series, Tags, Inside text).
-//! Mounted by `TopNav`; web-only via `cfg(not(feature = "mobile"))`.
-
-use dioxus::core::Task;
-use dioxus::prelude::*;
-use dioxus_router::use_navigator;
-use omnibus_shared::{
-    PaletteAuthorHit, PaletteBookHit, PaletteResults, PaletteSeriesHit, PaletteTagHit,
-};
-
-use crate::{data, use_server_url, Route};
-
-/// Whether the search palette overlay is open. Registered at the App level
-/// via `use_context_provider` so both the trigger button and the global
-/// `⌘K` shortcut can toggle it.
-//
-// Hydration safety: the palette starts closed (`PaletteOpen = false`) and
-// `SpOverlay` only renders when open, so SSR and WASM agree on initial DOM
-// — no hydration mismatch. The `⌘K` listener fires only under
-// `feature = "web"`.
-#[derive(Copy, Clone, PartialEq)]
-pub struct PaletteOpen(pub Signal<bool>);
-
-/// Top-level host: renders the trigger button and (when open) the overlay.
-/// Mount this in `TopNav` in place of the old `NavSearch`.
-//
-// Component tree:
-//   SearchPaletteHost          — mounted in TopNav, replaces NavSearch
-//   ├─ SpTriggerButton         — search button (icon + "Search" + ⌘K kbd hint)
-//   └─ (open) SpOverlay        — portal: scrim + panel
-//              ├─ SpInput       — autofocused serif italic input
-//              ├─ SpMeta        — "5 results · 18ms"
-//              ├─ SpResultsList — scrollable grouped results
-//              └─ SpFooter      — keyboard hints + "fts5 · ranked"
-#[component]
-pub fn SearchPaletteHost() -> Element {
-    let open = use_context::<PaletteOpen>();
-
-    // Hotkey lives at App scope (see `use_palette_global_shortcut`):
-    // `TopNav` re-mounts on every route, so registering here would leak
-    // a fresh listener per visit and flip the signal multiple times per
-    // press.
-
-    rsx! {
-        SpTriggerButton { open }
-        if open.0() {
-            SpOverlay { open }
-        }
-    }
-}
-
-// ── Trigger button ───────────────────────────────────────────────
-
-/// Pill-shaped search button in the nav: search icon + "Search" + ⌘K hint.
-#[component]
-fn SpTriggerButton(open: PaletteOpen) -> Element {
-    let mut open = open;
-    rsx! {
-        button {
-            class: "sp-trigger",
-            "data-testid": "search-trigger",
-            r#type: "button",
-            onclick: move |_| open.0.set(true),
-            // Search icon (SVG magnifying glass)
-            svg {
-                class: "sp-trigger-icon",
-                width: "15",
-                height: "15",
-                view_box: "0 0 24 24",
-                fill: "none",
-                stroke: "currentColor",
-                stroke_width: "2",
-                stroke_linecap: "round",
-                stroke_linejoin: "round",
-                circle { cx: "11", cy: "11", r: "8" }
-                line { x1: "21", y1: "21", x2: "16.65", y2: "16.65" }
-            }
-            span { "Search" }
-            kbd { class: "sp-trigger-kbd", "⌘K" }
-        }
-    }
-}
-
-// ── Overlay (scrim + panel) ──────────────────────────────────────
-
-/// Floating overlay: dark scrim + centered panel with input, results, footer.
-#[component]
-fn SpOverlay(open: PaletteOpen) -> Element {
-    let mut open = open;
-    let server_url = use_server_url();
-    let mut query = use_signal(String::new);
-    let mut results = use_signal(|| Option::<PaletteResults>::None);
-    let mut selected = use_signal(|| 0_usize);
-    let mut loading = use_signal(|| false);
-    // Tracks whether the user has driven selection with arrow keys this
-    // session. When false, pressing Enter navigates to the full-page
-    // search results instead of drilling into `selected`.
-    let mut has_navigated = use_signal(|| false);
-    // #126: handle of the in-flight debounce+RPC task so the next keystroke
-    // can `.cancel()` it before spawning a new one (instead of leaving N
-    // sleeping spawns to race past the debounce and burn pool connections).
-    let mut current_task = use_signal(|| Option::<Task>::None);
-    let nav = use_navigator();
-
-    // Build a flat list of selectable items for keyboard navigation.
-    let flat_items = use_memo(move || build_flat_items(&results.read()));
-
-    // Close the palette.
-    let mut close = move || {
-        open.0.set(false);
-    };
-
-    let on_keydown = make_keydown_handler(open, selected, has_navigated, flat_items, query, nav);
-
-    // Debounced search. Uses gloo_timers on web, tokio::time on server.
-    // Each keystroke cancels the prior task (debounce sleep + RPC) before
-    // spawning a new one — only one task is in flight at a time, so stale
-    // requests don't race past the debounce or hit SQLite.
-    let url = server_url.clone();
-    let mut spawn_search = move |v: String| {
-        if let Some(prev) = current_task.write().take() {
-            prev.cancel();
-        }
-        let url = url.clone();
-        let task = spawn(async move {
-            async_sleep_ms(150).await;
-            let trimmed = v.trim().to_string();
-            if trimmed.is_empty() {
-                results.set(None);
-                loading.set(false);
-                return;
-            }
-            loading.set(true);
-            match data::search_palette(&url, &trimmed).await {
-                Ok(r) => {
-                    selected.set(0);
-                    results.set(Some(r));
-                }
-                Err(_) => {
-                    results.set(None);
-                }
-            }
-            loading.set(false);
-        });
-        current_task.set(Some(task));
-    };
-
-    let res = results.read();
-    // Only project `selected` into row class names once the user has driven
-    // selection with arrow keys. Otherwise the first row would render with
-    // the "selected" highlight on every fresh query (since `selected`
-    // resets to 0 after each search response), and pressing ArrowDown
-    // would visually jump to index 1 instead of 0.
-    let is_loading = loading();
-
-    let total = res.as_ref().map(|r| r.total_count()).unwrap_or(0);
-    let duration = res.as_ref().map(|r| r.duration_ms).unwrap_or(0);
-
-    rsx! {
-        div {
-            class: "sp-scrim",
-            "data-testid": "sp-scrim",
-            onclick: move |_| close(),
-            tabindex: "-1",
-            div {
-                class: "sp-panel",
-                "data-testid": "sp-panel",
-                role: "dialog",
-                aria_label: "Search palette",
-                aria_modal: "true",
-                // Stop clicks inside the panel from closing the scrim.
-                onclick: move |evt| evt.stop_propagation(),
-                onkeydown: on_keydown,
-
-                // Input row
-                div { class: "sp-input-wrap",
-                    svg {
-                        class: "sp-input-icon",
-                        width: "18",
-                        height: "18",
-                        view_box: "0 0 24 24",
-                        fill: "none",
-                        stroke: "currentColor",
-                        stroke_width: "2",
-                        stroke_linecap: "round",
-                        stroke_linejoin: "round",
-                        circle { cx: "11", cy: "11", r: "8" }
-                        line { x1: "21", y1: "21", x2: "16.65", y2: "16.65" }
-                    }
-                    input {
-                        class: "sp-input",
-                        "data-testid": "sp-input",
-                        r#type: "text",
-                        placeholder: "Search books, authors, series, tags…",
-                        autofocus: true,
-                        // `autofocus` only fires on initial page load.
-                        // `onmounted` programmatically focuses the input
-                        // when the overlay is dynamically rendered (⌘K).
-                        // Uses `requestAnimationFrame` so the browser has
-                        // finished layout before we `.focus()` — without
-                        // that delay the focus call lands on an element
-                        // that isn't yet painted and the caret never
-                        // appears (this is the timing reason prior
-                        // attempts with `set_focus(true)`/`spawn` failed).
-                        onmounted: move |evt: MountedEvent| {
-                            focus_palette_input(&evt);
-                        },
-                        value: "{query}",
-                        oninput: move |evt| {
-                            let v = evt.value();
-                            query.set(v.clone());
-                            spawn_search(v);
-                            // Typing reverts to "Enter goes to /search" until
-                            // the user re-engages arrow-key navigation.
-                            has_navigated.set(false);
-                            selected.set(0);
-                        },
-                    }
-                    if is_loading {
-                        span { class: "sp-spinner", "…" }
-                    }
-                }
-
-                // Meta line
-                if res.is_some() {
-                    div { class: "sp-meta", "data-testid": "sp-result-count",
-                        "{total} result{plural(total)} · {duration}ms"
-                    }
-                }
-
-                SpResultsList {
-                    results,
-                    flat_items,
-                    selected,
-                    has_navigated,
-                    open,
-                }
-
-                // Footer
-                SpFooter {}
-            }
-        }
-    }
-}
-
-// ── Keyboard handler ─────────────────────────────────────────────
-
-/// Build the `onkeydown` event handler for the palette panel.
-fn make_keydown_handler(
-    mut open: PaletteOpen,
-    mut selected: Signal<usize>,
-    mut has_navigated: Signal<bool>,
-    flat_items: Memo<Vec<FlatItem>>,
-    query: Signal<String>,
-    nav: dioxus_router::Navigator,
-) -> impl FnMut(Event<KeyboardData>) {
-    move |evt: Event<KeyboardData>| {
-        let key = evt.key();
-        match key {
-            Key::Escape => {
-                evt.prevent_default();
-                open.0.set(false);
-            }
-            Key::ArrowDown => {
-                evt.prevent_default();
-                let len = flat_items.read().len();
-                if len > 0 {
-                    // First arrow press is a "begin navigation" gesture — it
-                    // should highlight the first item, not increment past it.
-                    // Subsequent presses wrap around as usual.
-                    if !has_navigated() {
-                        selected.set(0);
-                    } else {
-                        selected.set((selected() + 1) % len);
-                    }
-                    has_navigated.set(true);
-                }
-            }
-            Key::ArrowUp => {
-                evt.prevent_default();
-                let len = flat_items.read().len();
-                if len > 0 {
-                    // Symmetric with ArrowDown: first press lands on the
-                    // last item; subsequent presses walk backward.
-                    if !has_navigated() {
-                        selected.set(len - 1);
-                    } else {
-                        selected.set(if selected() == 0 {
-                            len - 1
-                        } else {
-                            selected() - 1
-                        });
-                    }
-                    has_navigated.set(true);
-                }
-            }
-            Key::Enter => {
-                evt.prevent_default();
-                let items = flat_items.read();
-                if has_navigated() {
-                    // Navigate to the selected result. Books/Authors/Series go
-                    // to their detail pages; Tags fall through to /search (no
-                    // per-tag detail page exists yet).
-                    if let Some(item) = items.get(selected()) {
-                        match item {
-                            FlatItem::Book { uuid, .. } => {
-                                nav.push(Route::BookDetail { uuid: uuid.clone() });
-                            }
-                            FlatItem::Author { id, .. } => {
-                                nav.push(Route::AuthorDetail { id: *id });
-                            }
-                            FlatItem::Series { id, .. } => {
-                                nav.push(Route::SeriesDetail { id: *id });
-                            }
-                            FlatItem::Tag { name, .. } => {
-                                nav.push(Route::Search {
-                                    query: facet_query("tag", name),
-                                });
-                            }
-                        }
-                        open.0.set(false);
-                    }
-                } else {
-                    let q = query().trim().to_string();
-                    if !q.is_empty() {
-                        nav.push(Route::Search { query: q });
-                        open.0.set(false);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-// ── Grouped results list ─────────────────────────────────────────
-
-/// Scrollable grouped result list rendered inside the palette panel.
-///
-/// Owns the per-group heading + row layout. Each click callback closes the
-/// palette and navigates to the appropriate detail page via `use_navigator`.
-#[component]
-fn SpResultsList(
-    results: Signal<Option<PaletteResults>>,
-    flat_items: Memo<Vec<FlatItem>>,
-    selected: Signal<usize>,
-    has_navigated: Signal<bool>,
-    open: PaletteOpen,
-) -> Element {
-    let mut open = open;
-    let nav = use_navigator();
-    let res = results.read();
-    let items = flat_items.read();
-    let sel = selected();
-    let has_nav = has_navigated();
-
-    rsx! {
-        div { class: "sp-results",
-            if let Some(ref r) = *res {
-                // Books
-                if !r.books.is_empty() {
-                    SpGroupHead { label: "Books", count: r.books.len() }
-                    for book in r.books.iter() {
-                        SpBookRow {
-                            key: "{book.id}",
-                            book: book.clone(),
-                            selected: has_nav && is_selected(&items, sel, &FlatItem::Book { uuid: book.uuid.clone(), title: book.title.clone() }),
-                            on_click: {
-                                let uuid = book.uuid.clone();
-                                move |_| {
-                                    nav.push(Route::BookDetail { uuid: uuid.clone() });
-                                    open.0.set(false);
-                                }
-                            },
-                        }
-                    }
-                }
-
-                // Authors
-                if !r.authors.is_empty() {
-                    SpGroupHead { label: "Authors", count: r.authors.len() }
-                    for author in r.authors.iter() {
-                        SpAuthorRow {
-                            key: "{author.id}",
-                            author: author.clone(),
-                            selected: has_nav && is_selected(&items, sel, &FlatItem::Author { id: author.id, name: author.name.clone() }),
-                            on_click: {
-                                let id = author.id;
-                                move |_| {
-                                    nav.push(Route::AuthorDetail { id });
-                                    open.0.set(false);
-                                }
-                            },
-                        }
-                    }
-                }
-
-                // Series
-                if !r.series.is_empty() {
-                    SpGroupHead { label: "Series", count: r.series.len() }
-                    for s in r.series.iter() {
-                        SpSeriesRow {
-                            key: "{s.id}",
-                            series: s.clone(),
-                            selected: has_nav && is_selected(&items, sel, &FlatItem::Series { id: s.id, name: s.name.clone() }),
-                            on_click: {
-                                let id = s.id;
-                                move |_| {
-                                    nav.push(Route::SeriesDetail { id });
-                                    open.0.set(false);
-                                }
-                            },
-                        }
-                    }
-                }
-
-                // Tags
-                if !r.tags.is_empty() {
-                    SpGroupHead { label: "Tags", count: r.tags.len() }
-                    for tag in r.tags.iter() {
-                        SpTagRow {
-                            key: "{tag.id}",
-                            tag: tag.clone(),
-                            selected: has_nav && is_selected(&items, sel, &FlatItem::Tag { id: tag.id, name: tag.name.clone() }),
-                            on_click: {
-                                let name = tag.name.clone();
-                                move |_| {
-                                    nav.push(Route::Search {
-                                        query: facet_query("tag", &name),
-                                    });
-                                    open.0.set(false);
-                                }
-                            },
-                        }
-                    }
-                }
-
-                // Inside text — placeholder
-                SpGroupHead { label: "Inside text", count: 0 }
-                div { class: "sp-coming-soon", "data-testid": "sp-coming-soon",
-                    "Coming soon"
-                }
-            }
-        }
-    }
-}
-
-// ── Result rows ──────────────────────────────────────────────────
-
-#[component]
-fn SpGroupHead(label: &'static str, count: usize) -> Element {
-    rsx! {
-        div { class: "sp-group-head label",
-            if count > 0 {
-                "{label} · {count}"
-            } else {
-                "{label}"
-            }
-        }
-    }
-}
-
-#[component]
-fn SpBookRow(book: PaletteBookHit, selected: bool, on_click: EventHandler<MouseEvent>) -> Element {
-    let sel_class = if selected {
-        "sp-row selected"
-    } else {
-        "sp-row"
-    };
-    let server_url = use_server_url();
-    let cover = if book.cover_url.is_some() {
-        let url = format!("{server_url}/api/thumbs/{}/sm", book.id);
-        rsx! {
-            img {
-                class: "sp-row-cover",
-                src: "{url}",
-                alt: "",
-                loading: "lazy",
-            }
-        }
-    } else {
-        // Accent-backed fallback with first letter
-        let initial = book
-            .title
-            .chars()
-            .next()
-            .unwrap_or('?')
-            .to_uppercase()
-            .to_string();
-        let bg = book.accent.as_deref().unwrap_or("var(--bg-2)");
-        rsx! {
-            div {
-                class: "sp-row-cover sp-row-cover-fallback",
-                style: "background: {bg};",
-                "{initial}"
-            }
-        }
-    };
-
-    let year = book.year.as_deref().unwrap_or("");
-    let formats: String = book.formats.join(" · ");
-
-    rsx! {
-        div {
-            class: "{sel_class}",
-            "data-testid": "sp-book-row",
-            onclick: move |evt| on_click.call(evt),
-            {cover}
-            div { class: "sp-row-body",
-                div { class: "sp-row-title", "{book.title}" }
-                div { class: "sp-row-sub", "{book.author_display}" }
-            }
-            if !year.is_empty() || !formats.is_empty() {
-                div { class: "sp-row-meta",
-                    if !year.is_empty() { span { "{year}" } }
-                    if !formats.is_empty() { span { "{formats}" } }
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SpAuthorRow(
-    author: PaletteAuthorHit,
-    selected: bool,
-    on_click: EventHandler<MouseEvent>,
-) -> Element {
-    let sel_class = if selected {
-        "sp-row selected"
-    } else {
-        "sp-row"
-    };
-    let initial = author
-        .name
-        .chars()
-        .next()
-        .unwrap_or('?')
-        .to_uppercase()
-        .to_string();
-
-    rsx! {
-        div {
-            class: "{sel_class}",
-            "data-testid": "sp-author-row",
-            onclick: move |evt| on_click.call(evt),
-            div { class: "sp-avatar", "{initial}" }
-            div { class: "sp-row-body",
-                div { class: "sp-row-title", "{author.name}" }
-                div { class: "sp-row-sub",
-                    "{author.book_count} book{plural(author.book_count as usize)}"
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SpSeriesRow(
-    series: PaletteSeriesHit,
-    selected: bool,
-    on_click: EventHandler<MouseEvent>,
-) -> Element {
-    let sel_class = if selected {
-        "sp-row selected"
-    } else {
-        "sp-row"
-    };
-
-    rsx! {
-        div {
-            class: "{sel_class}",
-            "data-testid": "sp-series-row",
-            onclick: move |evt| on_click.call(evt),
-            div { class: "sp-avatar", "S" }
-            div { class: "sp-row-body",
-                div { class: "sp-row-title", "{series.name}" }
-                div { class: "sp-row-sub",
-                    "{series.book_count} book{plural(series.book_count as usize)}"
-                    if let Some(ref author) = series.author_display {
-                        " · {author}"
-                    }
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SpTagRow(tag: PaletteTagHit, selected: bool, on_click: EventHandler<MouseEvent>) -> Element {
-    let sel_class = if selected {
-        "sp-row selected"
-    } else {
-        "sp-row"
-    };
-
-    rsx! {
-        div {
-            class: "{sel_class}",
-            "data-testid": "sp-tag-row",
-            onclick: move |evt| on_click.call(evt),
-            span { class: "sp-tag-chip", "# {tag.name}" }
-            div { class: "sp-row-body",
-                div { class: "sp-row-sub",
-                    "{tag.book_count} book{plural(tag.book_count as usize)}"
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SpFooter() -> Element {
-    rsx! {
-        div { class: "sp-footer",
-            div { class: "sp-footer-keys",
-                kbd { "↑↓" }
-                span { " navigate" }
-                kbd { "⏎" }
-                span { " open" }
-                kbd { "esc" }
-                span { " close" }
-            }
-            div { class: "sp-footer-engine",
-                "fts5 · ranked by relevance"
-            }
-        }
-    }
-}
-
-// ── Flat item model for keyboard nav ─────────────────────────────
-
-/// A single selectable item in the flat list used for arrow-key navigation.
-/// Books carry the stable `uuid` so the palette can build `/books/:uuid`
-/// URLs without a second round-trip — [`PaletteBookHit`] now includes
-/// the uuid alongside `id`.
-#[derive(Clone, Debug, PartialEq)]
-enum FlatItem {
-    Book { uuid: String, title: String },
-    Author { id: i64, name: String },
-    Series { id: i64, name: String },
-    Tag { id: i64, name: String },
-}
-
-/// Build a flat ordered list of all selectable items from the results.
-fn build_flat_items(results: &Option<PaletteResults>) -> Vec<FlatItem> {
-    let Some(r) = results else {
-        return Vec::new();
-    };
-    let mut items = Vec::new();
-    for b in &r.books {
-        items.push(FlatItem::Book {
-            uuid: b.uuid.clone(),
-            title: b.title.clone(),
-        });
-    }
-    for a in &r.authors {
-        items.push(FlatItem::Author {
-            id: a.id,
-            name: a.name.clone(),
-        });
-    }
-    for s in &r.series {
-        items.push(FlatItem::Series {
-            id: s.id,
-            name: s.name.clone(),
-        });
-    }
-    for t in &r.tags {
-        items.push(FlatItem::Tag {
-            id: t.id,
-            name: t.name.clone(),
-        });
-    }
-    items
-}
-
-/// Check if a given flat item matches the currently selected index.
-fn is_selected(items: &[FlatItem], selected_idx: usize, candidate: &FlatItem) -> bool {
-    items.get(selected_idx) == Some(candidate)
-}
-
-/// Simple English plural suffix.
-fn plural(n: usize) -> &'static str {
-    if n == 1 {
-        ""
-    } else {
-        "s"
-    }
-}
-
-/// Build a facet query string where every whitespace-separated word in
-/// `value` is prefixed with `prefix:`. This ensures `build_fts_match`
-/// routes each token to the correct FTS5 column filter instead of
-/// treating trailing words as free-text (e.g. `tag:Dark tag:academia`
-/// rather than `tag:Dark academia`).
-fn facet_query(prefix: &str, value: &str) -> String {
-    value
-        .split_whitespace()
-        .map(|w| format!("{prefix}:{w}"))
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-// ── Async sleep (platform-gated) ─────────────────────────────────
-
-/// Platform-gated async sleep. Web uses `gloo_timers`, server uses `tokio`.
-#[cfg(feature = "web")]
-async fn async_sleep_ms(ms: u32) {
-    gloo_timers::future::TimeoutFuture::new(ms).await;
-}
-
-#[cfg(all(not(feature = "web"), feature = "server"))]
-async fn async_sleep_ms(ms: u32) {
-    tokio::time::sleep(std::time::Duration::from_millis(ms as u64)).await;
-}
-
-// ── Autofocus on overlay mount (platform-gated) ──────────────────
-
-/// Focus the palette input after the browser has painted it.
-///
-/// Replaces a prior `document::eval` + raw JS `requestAnimationFrame` +
-/// `querySelector` workaround with a fully typed `web_sys` path.
-///
-/// The `requestAnimationFrame` indirection is still required. Calling
-/// `.focus()` synchronously inside `onmounted` (or from a `spawn`ed
-/// future) lands before the element is laid out, so the caret never
-/// appears. Scheduling the focus for the next frame matches what the
-/// old eval block did and lets the browser finish layout first.
-#[cfg(feature = "web")]
-fn focus_palette_input(evt: &MountedEvent) {
-    use dioxus::web::WebEventExt;
-    use wasm_bindgen::prelude::*;
-
-    let Some(element) = evt.try_as_web_event() else {
-        return;
-    };
-    let Some(window) = web_sys::window() else {
-        return;
-    };
-    // `Closure::once_into_js` hands the callback to JS and frees the Rust
-    // closure (and the captured `element`) after the browser fires it once,
-    // so repeatedly opening/closing the palette doesn't accumulate leaks.
-    let cb = Closure::once_into_js(move || {
-        if let Some(html_el) = element.dyn_ref::<web_sys::HtmlElement>() {
-            let _ = html_el.focus();
-        }
-    });
-    let _ = window.request_animation_frame(cb.unchecked_ref());
-}
-
-#[cfg(not(feature = "web"))]
-fn focus_palette_input(_evt: &MountedEvent) {
-    // No-op on SSR / native: the overlay only opens via web interaction.
-}
-
-// ── Global ⌘K shortcut (web only) ───────────────────────────────
-
-/// Register a global `keydown` listener that toggles the palette on `⌘K`
-/// (Mac) or `Ctrl+K` (other platforms).
-///
-/// Must be called from a component that mounts **exactly once** for the
-/// app's lifetime (i.e. `App`), not from a component that re-mounts on
-/// route changes. `use_hook` only guarantees "once per component
-/// instance", so a re-mounting host would accumulate listeners — each
-/// press would then toggle the signal N times and effectively no-op on
-/// the second route onward.
-#[cfg(feature = "web")]
-pub fn use_palette_global_shortcut() {
-    let mut open = use_context::<PaletteOpen>();
-    use_hook(move || {
-        use wasm_bindgen::prelude::*;
-
-        let closure = Closure::wrap(Box::new(move |evt: web_sys::KeyboardEvent| {
-            let is_cmd_k = (evt.meta_key() || evt.ctrl_key()) && evt.key() == "k";
-            if is_cmd_k {
-                evt.prevent_default();
-                let current = open.0();
-                open.0.set(!current);
-            }
-        }) as Box<dyn FnMut(_)>);
-
-        if let Some(window) = web_sys::window() {
-            let _ = window
-                .add_event_listener_with_callback("keydown", closure.as_ref().unchecked_ref());
-        }
-
-        // Leak the closure so it lives for the app lifetime. The shortcut
-        // is registered once (use_hook guarantees this) and never removed —
-        // acceptable for a single-page app.
-        closure.forget();
-    });
-}
-
-/// Non-web no-op so the App component can call this unconditionally on
-/// non-mobile targets (SSR + native) without dragging in `web_sys`.
-#[cfg(all(not(feature = "web"), not(feature = "mobile")))]
-pub fn use_palette_global_shortcut() {}
